mgadm: rename ndp to bgp unnumbered
